Stop truncating the reverse arc weights in the odd-wheel auxiliary graph
Odd-cycle separation runs a shortest path in the bipartite double cover: each
conflict (x,y) becomes a forward arc (x',y'') and a reverse arc (y'',x'), and a
path from v' to v'' is an odd closed walk. Both copies of an edge must therefore
carry the same length.

They did not. The forward arcs get icaActivity_ = 1001 - 1000*x, which is not
integral, and the reverse loop copied it through

    const size_t arcDist = spArcDist_[i2];

so every reverse arc was up to 1.0 shorter than its own forward twin -- and with
weights in [1, 1000] that is up to a 100% error on the cheapest arcs, the ones a
minimum-weight walk is most likely to use. A path was then cheaper the more
reverse arcs it took, which is a direction bias with no meaning in the problem.
spArcDist_ is a vector<double> and the forward path stores the exact value; this
is a typo, not a deliberate rounding.

The fix changes which cycles the search finds, so it is not bit-identical and
lands on its own. Sweeping all 237 replay fixtures with --check-validity:

  --rounds=1: oddHoles 1228 -> 1281 (+4.3%), cuts 1117 -> 1119, summed bound
  improvement 1086 -> 1088. 3 fixtures better, 2 worse, 232 identical.
  --rounds=4: oddHoles 1634 -> 1676, cuts 1488 -> 1451 -- fewer cuts, and yet
  summed bound improvement 5460 -> 5514 (+1.0%). 6 better, 2 worse, 229
  identical. Best momentum1 +72.0, trdtatl9220 +3.0, trdHardCA4500-8 +0.5;
  worst co-100 -21.5 and trdta99 -0.4.

The certificate passes on every one of the 1281 / 1676 wheels, all five certBad*
counters 0, so the extra cycles the corrected lengths expose are valid cuts and
not an artifact of a broken graph.

+++ src/CoinOddWheelSeparator.cpp
@@ -444,7 +444,12 @@ bool CoinOddWheelSeparator::prepareGraph(double startTime) {
             assert(spArcTo_[i2] >= icaCount_);
 #endif
             const size_t arcTo = spArcTo_[i2] - icaCount_;
-            const size_t arcDist = spArcDist_[i2];
+            /* double, not size_t: the weights are icaActivity_ = 1001 - 1000*x
+             * and are not integral. Truncating here gave the reverse copy
+             * (y'', x') of a conflict a length up to 1 shorter than its forward
+             * arc (x', y''), so the two directions of the same edge disagreed and
+             * a path was cheaper the more reverse arcs it used. */
+            const double arcDist = spArcDist_[i2];
 
             if(idxArc + 1 > spArcCap_) {
                 spArcCap_ *= 2;
